requirements.txt

google-generativeai>=0.5.0
numpy>=1.24.0
tqdm>=4.66.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0


